added changes

+++ src/backend/DeLong.Application/DTOs/InvoiceItems/InvoiceItemCreationDto.cs
@@ -2,8 +2,8 @@
 
 public class InvoiceItemCreationDto
 {
-    public int InvoiceId { get; set; }
-    public int ProductId { get; set; }
+    public long InvoiceId { get; set; }
+    public long ProductId { get; set; }
     public decimal Quantity { get; set; }
     public decimal UnitPrice { get; set; }
 }
