+++ src/backend/DeLong.Application/DTOs/InvoiceItems/InvoiceItemUpdateDto.cs
@@ -3,8 +3,8 @@
 public class InvoiceItemUpdateDto
 {
     public long Id { get; set; }
-    public int InvoiceId { get; set; }
-    public int ProductId { get; set; }
+    public long InvoiceId { get; set; }
+    public long ProductId { get; set; }
     public decimal Quantity { get; set; }
     public decimal UnitPrice { get; set; }
 }
