+++ src/backend/DeLong.Domain/Entities/InvoiceItem.cs
@@ -4,10 +4,10 @@ namespace DeLong.Domain.Entities;
 
 public class InvoiceItem : Auditable
 {
-    public int InvoiceId { get; set; }
+    public long InvoiceId { get; set; }
     public required Invoice Invoice { get; set; }
 
-    public int ProductId { get; set; }
+    public long ProductId { get; set; }
     public required Product Product { get; set; }
 
     public decimal Quantity { get; set; }
