+++ src/backend/DeLong.Domain/Entities/Stock.cs
@@ -4,10 +4,10 @@ namespace DeLong.Domain.Entities;
 
 public class Stock : Auditable // ombordagi qoldiq
 {
-    public int WarehouseId { get; set; }
+    public long WarehouseId { get; set; }
     public required Warehouse Warehouse { get; set; }
 
-    public int ProductId { get; set; }
+    public long ProductId { get; set; }
     public Product? Product { get; set; }
 
     public decimal Quantity { get; set; } // maxsulot soni
